requirements.txt
tqdm==4.42.1
numpy==1.22.4
opencv_python==4.6.0
matplotlib==3.5.2
vizdoom==1.1.7
torch==1.11.0
gym==0.21.0
gym_minigrid==1.0.1
PyYAML==6.0
tensorboard=2.0
ray=1.13.0
